build(ci): pin node to 14 for now
Travis uses the latest version of npm that corresponds to the specified
node version. node 15 uses npm 7, which seems to treat npm run --prefix
slightly differently in that the environment in which that command runs
is no longer in eg, backend/ or frontend/ .

As a result, npm run --prefix frontend eslint will not run the eslint
command found in frontend/package.json.

Work around this by pinning to node 14 for now